+++ src/pub/admin/install/index.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * Acuparse - AcuRite®‎ smartHUB and IP Camera Data Processing, Display, and Upload.
- * @copyright Copyright (C) 2015-2017 Maxwell Power
+ * Acuparse - AcuRite®‎ Access/smartHUB and IP Camera Data Processing, Display, and Upload.
+ * @copyright Copyright (C) 2015-2018 Maxwell Power
  * @author Maxwell Power <max@acuparse.com>
  * @link http://www.acuparse.com
  * @license AGPL-3.0+
@@ -30,7 +30,11 @@ require(dirname(dirname(dirname(__DIR__))) . '/inc/loader.php');
 // Process an upgrade
 if (isset($_GET['update']) && $installed === true) {
     if (isset($_GET['do'])) {
-        require 'updates.php';
+        $notes = '';
+        foreach (glob("scripts/*.php") as $filename)
+        {
+            include $filename;
+        }
         // Save the users config file
         $export = var_export($config, true);
         $export = str_ireplace('stdClass::__set_state', '(object)', $export);