+++ src/pub/admin/install/scripts/2_1.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * Acuparse - AcuRite®‎ smartHUB and IP Camera Data Processing, Display, and Upload.
- * @copyright Copyright (C) 2015-2017 Maxwell Power
+ * Acuparse - AcuRite®‎ Access/smartHUB and IP Camera Data Processing, Display, and Upload.
+ * @copyright Copyright (C) 2015-2018 Maxwell Power
  * @author Maxwell Power <max@acuparse.com>
  * @link http://www.acuparse.com
  * @license AGPL-3.0+
@@ -21,7 +21,7 @@
  */
 
 /**
- * File: src/pub/admin/install/updates.php
+ * File: src/pub/admin/install/scripts/2_1.php
  * Site Update Tasks
  */
 
@@ -71,5 +71,23 @@ switch ($config->version->app) {
         $config->version->app = '2.1.9';
         $config->site->hide_alternate = 'false';
         $notes .= '<li>' . $config->version->app . ' - ' . 'New option to hide alternate measurements</li>';
+
+    // Update from 2.1.9
+    case '2.1.9':
+        $config->version->app = '2.2.0';
+        $config->version->schema = '2.2';
+        mysqli_query($conn, "UPDATE `system` SET `value` = '2.2' WHERE `system`.`name` = 'schema';"); //Update Schema Version
+        $config->station->access_mac = '000000000000'; // Add Access MAC
+
+        // Fix MyAcuRite upload variables
+        $config->upload->myacurite->enabled === true ? $config->upload->myacurite->hub_enabled = true : $config->upload->myacurite->hub_enabled = false;
+        $config->upload->myacurite->enabled === true ? $config->upload->myacurite->access_enabled = true : $config->upload->myacurite->access_enabled = false;
+        $config->upload->myacurite->url === 'http://hubapi.myacurite.com' ? $config->upload->myacurite->hub_url = 'http://hubapi.myacurite.com' : $config->upload->myacurite->hub_url = 'http://hubapi.acuparse.com';
+        $config->upload->myacurite->url === 'http://hubapi.myacurite.com' ? $config->upload->myacurite->access_url = 'https://atlasapi.myacurite.com' : $config->upload->myacurite->hub_url = 'https://atlasapi.acuparse.com';
+        unset($config->upload->myacurite->enabled);
+        unset($config->upload->myacurite->url);
+        $config->upload->myacurite->hub_url = 'http://hubapi.myacurite.com';
+        $config->upload->myacurite->access_url = 'https://atlasapi.myacurite.com';
+
+        $notes .= '<li>' . $config->version->app . ' - ' . 'Support for the Acurite Access.<br> NOTICE: Apache rebuild required. See <a href="https://github.com/acuparse/acuparse/tree/master/docs/updates/from_2.1.md">docs/updates/from2_1.md</a></li>';
 }
-//$config->version->schema = '2.1';